requirements.txt
dotenv
pyswip
mysql-connector-python